Renaming some packages to avoid using underscores
